BaseUpdater Class
This class is to share functionality for writing commands to import data
to the uk_geo_utils models, with minimal downtime.

The basic pattern is to:
  - create a temp table
  - use 'copy' to populate the temp table
  - build indexes on the temp table
  - within a transaction rename the temp table to replace the existing
    table.

This is made more complicated because sometimes we have to deal with
primary keys, foreign keys and making sure all the names are the same
before and after.