Move to August 2024 ONSPD header
This commit also brings in checks to let us know if the header is
correct. If it's not then we get a list of missing/extra fields.
This means it should be obvious next time they change it.

It will also help us catch non erroring changes like nuts -> itl
which we had missed.