Create temp tables unlogged
This stops anything to do with the table being written to the WAL or
being replicated. This is good because we have publication set to all
tables on WDIV and so when we created the temp table it would try and
replicate it, then hit the following error:

> OperationalError: cannot update table "uk_geo_utils_onspd_temp"
> because it does not have a replica identity and publishes updates
> HINT:  To enable updating the table, set REPLICA IDENTITY using ALTER TABLE.

This stops that happening.