Alter temp table replica identity during import.
Before we import data set replica identity to 'FULL'.
After importing data and building indexes set it back to 'DEFAULT'.

This means wal files on the temp tables are written using the full row,
which probably takes up more storage, but it's only for the duration of
the import.

It allows us to keep an 'alltables' publication, rather than
having to curate a list of the tables we want to replicate.

The replication identity is carried with the temp table when it is
renamed to replace the old table.

However this seems to break replication fairly reliably. On The RDS `select * from pg_replication_slots;` returns a slot with `active: f`, ``wal_status: lost` and `conflicting: t`.
So even with this 'fix' it will be necessary to cycle the instances after an addressbase update to get the local dbs in sync with the RDS. Thankfully this seems to do the trick.